Move file into _geoarrow dir
