+++ lonboard/_geoarrow/_duckdb.py
